+++ .github/pull_request_template.md
@@ -3,7 +3,8 @@
 3. Deploy: [link](https://github.com/)
 4. Done 28.05.2020 / deadline 31.05.2020
 5. Score: 75 / 110
-  - [x] Application is divided to logical modules / layers. Like working with api it's separate module, as well as working with UI rendering, or working with an application state, etc. Better discuss architecture with a mentor before implementing. (40/40)
-  - [x] All the HTML-content is generated by JavaScript (body contains only one tag script inside). (15/30) 
-  - [ ] Application is SPA.(0/20)
-  - [x] Webpack or another bundler should be used. There should be one HTML-file, JS-file, and one CSS-file (or it can be embedded right in the HTML-file). (20/20)
+
+- [x] Application is divided to logical modules / layers. Like working with api it's separate module, as well as working with UI rendering, or working with an application state, etc. Better discuss architecture with a mentor before implementing. (40/40)
+- [x] All the HTML-content is generated by JavaScript (body contains only one tag script inside). (15/30)
+- [ ] Application is SPA.(0/20)
+- [x] Webpack or another bundler should be used. There should be one HTML-file, JS-file, and one CSS-file (or it can be embedded right in the HTML-file). (20/20)
