+++ .husky/pre-commit
@@ -1,0 +1,3 @@
+echo "Husky pre-commit start"
+npm run lint
+echo "Husky pre-commit finished!"
