--- class-components/.husky/pre-commit
@@ -1,1 +1,0 @@
-npm run lint
