--- class-components/src/components/SearchButton/SearchButton.tsx
@@ -1,3 +1,0 @@
-export default function SearchButton() {
-    return <button onClick={(e) => e.preventDefault()}/>;
-}
