+++ src/components/SearchButton/SearchButton.tsx
@@ -1,0 +1,3 @@
+export default function SearchButton() {
+  return <button onClick={(e) => e.preventDefault()} />;
+}
